LiveAPI: Work done on proto specifications, event mappers, added info/error support
