+++ events/eventmessages/adminBlock.proto
@@ -1,8 +1,10 @@
 syntax = "proto3";
 package eventmessages;
+option go_package = "eventmessages";
+option java_package = "com.factom.factomd.eventmessages";
 
-import "common/messages/eventmessages/generalTypes.proto";
-import "common/messages/eventmessages/factoidBlock.proto";
+import "events/eventmessages/generalTypes.proto";
+import "events/eventmessages/factoidBlock.proto";
 import "google/protobuf/timestamp.proto";
