+++ events/eventmessages/factomEvents.proto
@@ -1,10 +1,12 @@
 syntax = "proto3";
 package eventmessages;
+option go_package = "eventmessages";
+option java_package = "com.factom.factomd.eventmessages";
 
-import "common/messages/eventmessages/generalTypes.proto";
-import "common/messages/eventmessages/adminBlock.proto";
-import "common/messages/eventmessages/factoidBlock.proto";
-import "common/messages/eventmessages/entryCredit.proto";
+import "events/eventmessages/generalTypes.proto";
+import "events/eventmessages/adminBlock.proto";
+import "events/eventmessages/factoidBlock.proto";
+import "events/eventmessages/entryCredit.proto";
 import "google/protobuf/timestamp.proto";
 
 
@@ -12,7 +14,8 @@ enum EventSource {
     COMMIT_DIRECTORY_BLOCK = 0;
     ADD_TO_PROCESSLIST = 1;
     ADD_TO_HOLDING = 2;
-    NODE_ERROR = 3;
+    NODE_INFO = 3;
+    NODE_ERROR = 4;
 }
 
 
@@ -23,7 +26,7 @@ message FactomEvent {
         CommitChain commitChain = 3;
         CommitEntry commitEntry = 4;
         RevealEntry revealEntry = 5;
-        Error error = 6;
+        string nodeMessage = 6;
     }
 }
