+++ events/eventmessages/generalTypes.proto
@@ -1,5 +1,7 @@
 syntax = "proto3";
 package eventmessages;
+option go_package = "eventmessages";
+option java_package = "com.factom.factomd.eventmessages";
 
 message Hash {
     bytes hashValue = 1;
@@ -21,10 +23,4 @@ message Signature {
 message TransAddress {
     uint64 amount = 1;
     Hash address = 2;
-}
-
-
-message Error {
-    string message = 1;
-    string source = 2;
 }
